requirements.txt
openai==1.56.1
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
requests==2.31.0
firebase-admin==6.5.0
backoff==2.2.1
pinecone==7.3.0